+++ web/store/state.store.ts
@@ -91,9 +91,8 @@ export class StateStore implements IStateStore {
   createState = async (workspaceSlug: string, projectId: string, data: Partial<IState>) => {
     const response = await this.stateService.createState(workspaceSlug, projectId, data);
 
-    const _stateMap = set(this.stateMap, [response?.id], response);
     runInAction(() => {
-      this.stateMap = _stateMap;
+      set(this.stateMap, [response?.id], response);
     });
     return response;
   };
@@ -109,9 +108,8 @@ export class StateStore implements IStateStore {
   updateState = async (workspaceSlug: string, projectId: string, stateId: string, data: Partial<IState>) => {
     const originalState = this.stateMap[stateId];
     try {
-      const _stateMap = set(this.stateMap, [stateId], { ...this.stateMap?.[stateId], ...data });
       runInAction(() => {
-        this.stateMap = _stateMap;
+        set(this.stateMap, [stateId], { ...this.stateMap?.[stateId], ...data });
       });
       const response = await this.stateService.patchState(workspaceSlug, projectId, stateId, data);
       return response;
@@ -135,12 +133,12 @@ export class StateStore implements IStateStore {
   deleteState = async (workspaceSlug: string, projectId: string, stateId: string) => {
     const originalStates = this.stateMap;
     try {
-      const _stateMap = this.stateMap;
-      delete this.stateMap[stateId];
+      if (!this.stateMap?.[stateId]) return;
 
       runInAction(() => {
-        this.stateMap = _stateMap;
+        delete this.stateMap[stateId];
       });
+
       await this.stateService.deleteState(workspaceSlug, projectId, stateId);
     } catch (error) {
       runInAction(() => {
@@ -159,9 +157,8 @@ export class StateStore implements IStateStore {
   markStateAsDefault = async (workspaceSlug: string, projectId: string, stateId: string) => {
     const originalStates = this.stateMap;
     try {
-      const _stateMap = set(this.stateMap, [stateId, "default"], true);
       runInAction(() => {
-        this.stateMap = _stateMap;
+        set(this.stateMap, [stateId, "default"], true);
       });
 
       await this.stateService.markDefault(workspaceSlug, projectId, stateId);
@@ -204,9 +201,8 @@ export class StateStore implements IStateStore {
         else newSequence = (groupStates[groupIndex + 2].sequence + groupStates[groupIndex + 1].sequence) / 2;
       }
 
-      const _stateMap = set(this.stateMap, [stateId, "sequence"], newSequence);
       runInAction(() => {
-        this.stateMap = _stateMap;
+        set(this.stateMap, [stateId, "sequence"], newSequence);
       });
 
       // updating using api
